google fill in the blank

+++ python/board_test.py
@@ -22,48 +22,24 @@ if __name__ == '__main__':
     # clues_file = os.path.join(root, 'Oct2219-clues.txt')
     # entries_file = os.path.join(root, 'Oct2219-entries.txt')
     im = imageio.imread(os.path.join(root, 'img_test', impath))
-
     board = make_board(im)
-    if False:
-        # answers_file = os.path.join(root, 'smogon-answers-partial.txt')
-        answers_file = os.path.join(root, 'smogon-answers-minus.txt')
-        # answers_file = os.path.join(root, 'smogon-answers.txt')
-        all_answers = defaultdict(lambda: [None]) # type: DefaultDict[int, List[Optional[str]]]
-        for line in open(answers_file):
-            for word in line.strip().split():
-                if word.isalpha():
-                    all_answers[len(word)].append(word.upper())
 
-        for entries in board.entries:
-            for entry in entries:
-                answers = all_answers[len(entry)]
-                scores = [1] * len(answers)
-                scores[0] = 10
-                entry.set_answers(answers, scores)
-
-    with open(clues_file) as f:
-        clues = f.read()
-    # if os.path.exists(entries_file):
-        # with open(entries_file) as f:
-            # board.load_entries(f.read(), weight_for_unknown=0)
-    # else:
-        # board.use_clues(clues, weight_for_unknown=10)
-        # with open(entries_file, 'w') as f:
-            # f.write(board.dump_entries())
-    board.use_clues(clues, weight_for_unknown=10)
-    with open(entries_file, 'w') as f:
-        f.write(board.dump_entries())
+    weight_for_unknown = 10
+    force_recompile = True
+    if not force_recompile and os.path.exists(entries_file):
+        with open(entries_file) as f:
+            board.load_entries(f.read(), weight_for_unknown=weight_for_unknown)
+    else:
+        with open(clues_file) as f:
+            clues = f.read()
+        board.use_clues(clues, weight_for_unknown=weight_for_unknown)
+        with open(entries_file, 'w') as f:
+            f.write(board.dump_entries())
 
     for i in range(30):
         board.update_cells()
         board.update_entries()
     board.update_cells()
 
-    # output = board.format_multiple()
-    output = board.format(fill=True, number=False, probabilities=False)
-    # output = board.format(fill=False, number=True, probabilities=False)
-    # output = board.format(fill=True, number=False, probabilities=False, reduce_op=operator.itemgetter(1))
-    # output = board.format(fill=False, number=False, probabilities=True)
-    # output = board.format(fill=False, number=False, probabilities=True, reduce_op=operator.itemgetter(0))
-    # output = board.format(fill=False, number=False, probabilities=True, reduce_op=functools.partial(np.max, axis=0))
+    output = board.format_multiple()
     set_clipboard(html=output)
